Rename tasmota base entity module
