+++ homeassistant/components/tasmota/entity.py
