+++ homeassistant/components/tasmota/fan.py
@@ -24,7 +24,7 @@ from homeassistant.util.percentage import (
 
 from .const import DATA_REMOVE_DISCOVER_COMPONENT
 from .discovery import TASMOTA_DISCOVERY_ENTITY_NEW
-from .mixins import TasmotaAvailability, TasmotaDiscoveryUpdate
+from .entity import TasmotaAvailability, TasmotaDiscoveryUpdate
 
 ORDERED_NAMED_FAN_SPEEDS = [
     tasmota_const.FAN_SPEED_LOW,
