+++ homeassistant/components/tasmota/switch.py
@@ -15,7 +15,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import DATA_REMOVE_DISCOVER_COMPONENT
 from .discovery import TASMOTA_DISCOVERY_ENTITY_NEW
-from .mixins import TasmotaAvailability, TasmotaDiscoveryUpdate, TasmotaOnOffEntity
+from .entity import TasmotaAvailability, TasmotaDiscoveryUpdate, TasmotaOnOffEntity
 
 
 async def async_setup_entry(
